requirements.txt
# Core Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
selenium>=4.15.0
python-dotenv>=1.0.0

# API Integrations
openai>=1.3.0
notion-client>=2.2.0

# Utilities
ratelimit>=2.2.1
python-dateutil>=2.8.2

# Email
secure-smtplib>=0.1.1
